Client/Server project for visitors control system
The server will be multithreaded. The main server thread takes care of
the connections requests (establish rendezvous); the spawned
server-threads will carry-out the extended rendezvous with the clients.
The clock thread will be on the server site. The main server thread will
create the clock thread once main server starts. The clock thread can
control the number of main server connections, (6 visitors and 1 speaker
for each session). After the session, clock can close the connection for
some period (break) then open up again after the break to allow the next
group of 6 visitors and speaker to reconnect to the main server.
On the client side, we will have multiple client-threads (visitors,
speaker) that
execute concurrently. The client will ask the main server thread to
establish a
connection.
When the connection is accepted by the server, the main server will
create another
server-thread that will carry out the two-way communication with the
client thread. On the client side, we will have multiple client-threads
(visitors, speaker) that execute concurrently. The client will ask the
main server thread to establish a connection.
When the connection is accepted by the server, the main server will
create another server-thread that will carry out the two-way
communication with the client thread.